requirements.txt
Flask==3.0.0
pdfminer.six==20231228
docx2txt==0.8
